Update patch set 7
Patch Set 7:

The Go 1 compatibility guarantees permit us to add a new field to MarshalerError.  https://golang.org/doc/go1compat says "For the addition of features in later point releases, it may be necessary to add fields to exported structs in the API."  So I think we can just add a new field to MarshalerError, which seems simpler.  Perhaps if the new field is not set MarshalerError.Error should report "MarshalJSON".

Patch-set: 7
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>