Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > One MB is clearly too small; valid object files can easily have a larger string table.  I would believe one GB.
> 
> I agree. How did you pick 1 MB?
> 
> Alex

I picked 1MB based on analysis of millions of PE files in the wild and a specific concern for code which may be being run in relatively low memory environments. I have yet to have observed a string table over 1MB in size, but that is ultimately just anecdotal. I agree it is reasonably possible to have a larger string table, 1GB does seems a bit large to me though, considering that is above the size of almost all PE files.

I did originally think about trying to limit the string table size to the total size of the PE file (or section) being analysed, but that data seemed quite far out of context by the time we get to the readStringTable function.

Niall

Patch-set: 1
CC: Gerrit User 34263 <34263@62eb7196-b449-3ce5-99f1-c037f21e1705>