cmd/compile: report single error for invalid complex const conversions
Modifies toint to return a boolean when conversions fail, but
purposefully makes it return the empty Value{} and false
when we try to convert a complex value to an *int.

Given:
    _ = uint(-4 + 2i)

the new error will be
    cannot convert -4 + 2i (type untyped complex) to type uint

Perhaps we should also apply the same suppressing logic to
other const conversions, but we can think about that in
follow-up CLs.

Fixes #33921

Change-Id: Idb00b2e56468b41b07911f746f22da512faccace