Update patch set 2
Patch Set 2:

> Patch Set 2: TryBot-Result-1
> 
> 4 of 37 TryBots failed:
> Failed on linux-ppc64-buildlet: https://storage.googleapis.com/go-build-log/a6426def/linux-ppc64-buildlet_4b1773c6.log

Hmm. I still have a deadlock. I think this is happening now because a suspended G actually parks, so if it's the last G on the M, the M will park. I don't understand why the suspending M doesn't count toward checkdead, though.

> Failed on darwin-386-10_14: https://storage.googleapis.com/go-build-log/a6426def/darwin-386-10_14_6fbf3d45.log

This one's legit. I have a fix that I tested on a gomote.

> Failed on netbsd-amd64-8_0: https://storage.googleapis.com/go-build-log/a6426def/netbsd-amd64-8_0_8411a8c6.log
> Failed on netbsd-386-8_0: https://storage.googleapis.com/go-build-log/a6426def/netbsd-386-8_0_21aa8402.log

Same failures as an earlier CL in this series. I'll try on a parent and see if I introduced this.

Patch-set: 2
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>