runtime: make the scavenger self-pacing
Currently the runtime background scavenger is paced externally,
controlled by a collection of variables which together describe a line
that we'd like to stay under.

However, the line to stay under is computed as a function of the number
of free and unscavenged huge pages in the heap at the end of the last
GC. Aside from this number being inaccurate (which is still acceptable),
the scavenging system also makes an order-of-magnitude assumption as to
how expensive scavenging a single page actually is.

This change simplifies the scavenger in preparation for making it
operate on bitmaps. It makes it so that the scavenger paces itself, by
measuring the amount of time it takes to scavenge a single page. The
scavenging methods on mheap already avoid breaking huge pages, so if we
scavenge a real huge page, then we'll have paced correctly, otherwise
we'll sleep for longer to avoid using more than scavengePercent CPU
time.

Unfortunately, all this involves measuring time, which is quite tricky.
Currently we don't account for long process sleeps or OS-level context
switches, but we do account for scheduler overhead and variations in it
by maintaining an EWMA which represents how much time we actually spent
scavenging vs. sleeping, and adjusting the sleep time accordingly.

Change-Id: Ieca8b088fdfca2bebb06bcde25ef14a42fd5216b