Update patch set 5
Patch Set 5:

> Patch Set 5: -Run-TryBot -Code-Review
> 
> Seems like the base64 import violates the policy defined in go/build/deps_test.go. Is it absolutely needed for Unmarshal? If so, this needs to be discussed in the issue.

Hey Akhil,

This actually was discussed on the issue. The claim was that in order to be truly backwards compatible with pervious versions of Go, we needed to be able to decode the potentially base64 encoded byte arrays that theoretically may be have serialized in current and previous versions of Go. I don't know of any other way to do that besides using the base64 package (or copy and pasting from the base64 package). It seems we're stuck between two requirements here. Which would you prefer I go with? Or is there a third option?

Patch-set: 5