runtime: work around bad first sysReserve from wasm
When adding the new page allocator, I found that the first few bytes of
the first reservation (which is actually done by the new page
allocation) made by sysReserve ends up not being completely zero.
However, it seems the rest of the reservation is indeed zero. Work
around this problem by rounding up the initial reservation pointer to
the next page.

Change-Id: I665937f4c8e1494d4c70329e92c6f94340085c03