cmd/compile, cmd/link: add coverage instrumentation for fuzzing
DO NOT REVIEW
DO NOT SUBMIT

This CL adds basic coverage instrumentation similar to what
github.com/dvyukov/go-fuzz produces in its -libfuzzer mode. The
coverage can be enabled by compiling with -d=fuzzing. It's intended to
be used in conjunction with -buildmode=c-archive to produce an ELF .a
file that can be linked with libFuzzer.

The coverage generates a unique 8-bit counter for each basic block in
the original source code, and emits an increment operation. These
counters are then collected into the __libfuzzer_extra_counters ELF
section for use by libFuzzer.

Updates #14565.

Change-Id: I239758cc0ceb9ca1220f2d9d3d23b9e761db9bf1