archive/zip: support adding raw files
The existing implementation provides no way to add files to an archive where the
caller already knows the uncompressed size, compressed size and crc32 value of
the data to be written.

This commit introduces a CreateHeaderRaw function that takes the already known
information that can be set on the FileHeader fields and allows the caller to
write data that doesn't pass through any compressor. This allows for adding
already compressed data to the archive, without incurring the associated
compression overheads.

Fixes #34974

Change-Id: Ia13ebbdab725e3912a979d6ef510e7236129810f