Update patch set 34
Patch Set 34:

> SlowBots beginning. Status page: https://farmer.golang.org/try?commit=dbad0500

This is hanging waiting for the Solaris builder to become ready, so I'm just going to submit and carry on.  As far as I know the Solaris code is fine.  If it is not, I'll fix it later.

Patch-set: 34
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>