Update patch set 1
Patch Set 1:

The description says that you are removing unnecessary boundary checks, but that's not what you are doing.  You are changing the behavior of these methods when a short slice is passed in.

I don't see why that is desirable.  The methods expect full size slices and should not be called on short slices.  It's very easy for calling code to avoid calling these methods if they will not work.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>