runtime/internal/atomic: shorten s390x assembly for And8 and Or8
By using a 'rotate then xor selected bits' instruction combined with
either a 'load and and' or a 'load and or' instruction we can
implement And8 and Or8 with far fewer instructions. And8 and Or8
are intrinsics on other platforms and I'd like to make them
intrinsics on s390x. A shorter sequence of instructions will help
make that simpler. Replacing 'compare and swap' with atomic
instructions may also improve performance when there is contention.

Change-Id: I9ba0144f792fa8a7626f75bb92f4bd4648d23c75