syscall/js: garbage collect references to JavaScript values
The js.Value struct now also contains a pointer, so a finalizer can
determine if the value is not referenced by Go any more.

Unfortunately this breaks Go's == operator with js.Value. This change
also adds a new Equals method to check for the equality of two js.Value.
This is a breaking change, but its impact should be limited since
the == operator still works fine with js.Undefined() and js.Null().

Fixes #35111

Change-Id: I58a50ca18f477bf51a259c668a8ba15bfa76c955