Revert "cmd/link: use libmsvcrt.a during internal link"
This reverts CL 197977.

Reason for revert: as discussed in the issue, this seems to be a problem of a broken gcc, not with something we're doing. Rather than confusing symbols from different libraries, we should instead be using a mingw that contains the right symbols. Otherwise mayhem may ensue.

Updates #23649
Updates #35006

Change-Id: I4a5e5b470cd0ae70642e16177c75542830602944