runtime: add async preemption support on PPC64
This CL adds support of call injection and async preemption on
PPC64.

For the injected call to return to the preempted PC, we have to
clobber either LR or CTR. For reasons mentioned in previous CLs,
we choose CTR. Previous CLs have marked code sequences that use
CTR async-nonpreemtible.

DO NOT SUBMIT. ppc64le seems to work reliably, but there is still
some problem on big-endian ppc64.

Change-Id: Ia642b5f06a890dd52476f45023b2a830c522eee0