Update patch set 1
Patch Set 1:

(4 comments)

Thanks for doing this. I've suggested a few additional instructions that we should also include when adding these.

In general our coverage of the vector instructions is not great. We should add a lot more. This is a good start though and we can add more during the early part of the freeze since the code we are adding to is a test.

Patch-set: 1
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>