cmd/internal/obj/x86: prevent jumps crossing 32 byte boundaries
The microcode update for the Jump Conditional Code Erratum
may impact the performance of some workloads on some Intel
CPUs:

https://www.intel.com/content/www/us/en/support/articles/000055650.html

This commit mitigates against these performance issues by
adding a new option to the x86 assembler.  If the
GO_X86_PADJUMP environment variable is set to 1 and we're
doing a 64 bit build, the assembler will make sure that
neither stand-alone nor macro-fused jumps will end on or
cross 32 byte boundaries.  To achieve this, functions are
aligned on 32 byte boundaries, rather than 16 bytes, and
jump instructions are padded to ensure that they do not
cross or end on 32 byte boundaries.  Two padding strategies
are adopted:

1. The preferred strategy is to find instructions that occur
   before the jump that requires padding and then to insert
   legacy segment prefixes to start of these instructions.
   The prefixes are ignored by the CPU.
2. There may not always be a suitable instruction available
   into which prefixes can be inserted.  In this case we
   add a NOP instruction of the appropriate length before
   the jump.

For more information about how this patch works and why it
is needed, the reader is referred to the white paper
linked above.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>

Change-Id: Ib350138660727b5958289543f747e70f66616081