Update patch set 4
Patch Set 4:

I've rebased and fixed the conflicts once more. Seebs, do you want to add your +1 once more? It's okay if you've gotten bored of it :)

Brad/Joe, any chance of a review for 1.14? I realise that this patch is a bit non-trivial, but remember that most of the changes are just code being moved around. This is possible to see with 'git diff' or other tools.

I want to keep the pestering to a minimum, but you two are listed as owners alongside me :)

Patch-set: 4
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>